requirements.txt
streamlit>=1.30.0
arxiv>=2.1.0
tenacity>=8.2.3
PyMuPDF>=1.23.0
PyYAML>=6.0.1
requests>=2.31.0
openai>=1.0.0
anthropic>=0.25.0
